Removed Apache::Tika dependency
The code base for Apache::Tika provides no
error handling or ability to handle errors
as it silently squashes them.

To work around this, this change replaces it
with equivalent LWP::UserAgent code that calls
just the interface we need, and handles the
case where the server goes away or a document
doesn't process.

Also updated the README to correct some issues
with the installation and running instructions.